% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal.R
\name{build_coal_tree}
\alias{build_coal_tree}
\title{Generate a Newick string representation of a tree corresponding to a realisation of a coalescent process.
Note: the topology is RANDOM, If you wish for the topology to be consistent, use set.seed before calling this function.}
\usage{
build_coal_tree(samp_t, n_samp, coal_t, node_name_prefix = "N")
}
\arguments{
\item{samp_t}{sampling times.}

\item{coal_t}{times of coalescent events / internal nodes.}
}
\value{
an ape phylogeny with appropriate times
}
\description{
Generate a Newick string representation of a tree corresponding to a realisation of a coalescent process.
Note: the topology is RANDOM, If you wish for the topology to be consistent, use set.seed before calling this function.
}
