% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer_costs2}
\alias{infer_costs2}
\title{Run fitness cost inference}
\usage{
infer_costs2(
  phys,
  mrsts,
  ABX_usage,
  ABX_times,
  t_begin,
  t_end,
  gamma_guess,
  time_scale_d,
  gamma_log_sd = 0.1,
  n_iter = 2000,
  n_warmup = 2000,
  model = "deterministic",
  L = 6.5,
  K = 65,
  seed = 1,
  max_dt = 0.001,
  stan_control = stan_defaults()
)
}
\arguments{
\item{phys}{A list of phylogenies. First phylogeny is assumed to susceptible}

\item{mrsts}{A list of most recent sampling times (in coalescent time) relative to t_begin}

\item{ABX_usage}{AMR usage values as a proportion of primary treatment}

\item{ABX_times}{AMR usage time points}

\item{t_begin}{Analysis start date in natural time}

\item{t_end}{Analysis end date in natural time}

\item{gamma_guess}{Average recovery rate per day}

\item{time_scale_d}{The scale of one time unit for phylogenies and AMR data, in days}

\item{gamma_log_sd}{Log sd for strongly informative log-normal prior on average recovery rates}

\item{n_iter}{Number of sampling iterations}

\item{n_warmup}{Number of warmup iterations#}

\item{L}{HSGP approximation boundary factor}

\item{K}{HSGP approximation basis function count}

\item{seed}{RNG seed used by sampler}

\item{max_dt}{Maximum interval width for likelihood approximation on [-1,1] scale}

\item{stan_control}{List of additional arguments to pass to cmdstan. Accepted keywords: 'adapt_delta', 'max_treedepth', 'refresh', 'init', 'chains', 'parallel_chains'. See stan_defaults() for default values.}
}
\description{
Run fitness cost inference
}
