% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_epi_dynamic}
\alias{plot_epi_dynamic}
\title{Plot epidemic trajectory credible intervals and posterior draws}
\usage{
plot_epi_dynamic(o, lineage_index, n_draws = 40, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{lineage_index}{index of lineage to display, '1' indicates susceptible}

\item{n_draws}{Number of posterior epidemic trajectory draws to sample}

\item{...}{Additional parameters}
}
\description{
Plot epidemic trajectory credible intervals and posterior draws
}
