% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_epi_pairs}
\alias{plot_epi_pairs}
\title{Plot posterior marginals and scatter plots for recovery rates using bayesplot}
\usage{
plot_epi_pairs(o, res_lineage_idx, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{res_lineage_idx}{index of resistant lineage to display}

\item{...}{Additional parameters}
}
\description{
Plot posterior marginals and scatter plots for recovery rates using bayesplot
}
