% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_hyperpar_pairs}
\alias{plot_hyperpar_pairs}
\title{Plot posterior marginals for GP hyperparameters}
\usage{
plot_hyperpar_pairs(o, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{...}{Additional parameters}
}
\description{
Plot posterior marginals for GP hyperparameters
}
