% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_Ne}
\alias{plot_Ne}
\title{Plot Ne trajectories credible intervals and posterior draws}
\usage{
plot_Ne(o, lineage_index, n_draws = 40, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{lineage_index}{index of lineage to display, '1' indicates susceptible}

\item{n_draws}{Number of posterior Ne trajectory draws to sample}

\item{...}{Additional parameters}
}
\description{
Plot Ne trajectories credible intervals and posterior draws
}
