% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_ppcheck_At}
\alias{plot_ppcheck_At}
\title{Plot posterior predictive draws and data overlay for the lineages through time function}
\usage{
plot_ppcheck_At(o, lineage_index, n_draws = 40, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{lineage_index}{index of lineage to display, '1' indicates susceptible}

\item{n_draws}{number of trajectories to simulate and overlay}
}
\description{
Plot posterior predictive draws and data overlay for the lineages through time function
}
