% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_qpairs}
\alias{plot_qpairs}
\title{Plot q parameter marginals for all resistant strains}
\usage{
plot_qpairs(o, ...)
}
\arguments{
\item{o}{costsFit2 object}
}
\description{
Plot q parameter marginals for all resistant strains
}
