% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_rr_curve}
\alias{plot_rr_curve}
\title{Plot Rt_res / Rt_sus ratio posterior probability curve for a given threshold}
\usage{
plot_rr_curve(o, res_lineage_idx, rr_threshold = 0, n_breaks = 100)
}
\arguments{
\item{o}{costsFit2 object}

\item{res_lineage_idx}{index of resistant lineage to display}

\item{rr_threshold}{r_t ratio for which to plot the curve}

\item{n_breaks}{number of usage discretisation breaks}
}
\description{
f(U) = P[rt_sus - rt_res > c | u(t) = U]
}
