% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_rt}
\alias{plot_rt}
\title{Plot growth rate r(t) credible intervals and posterior draws}
\usage{
plot_rt(o, lineage_index, n_draws = 40, ...)
}
\arguments{
\item{o}{costsFit2 object}

\item{lineage_index}{index of lineage to display, '1' indicates susceptible}

\item{n_draws}{Number of posterior r(t) draws to sample}
}
\description{
Plot growth rate r(t) credible intervals and posterior draws
}
