% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot traces for a subset of parameters using bayesplot}
\usage{
plot_traces(o, ...)
}
\arguments{
\item{o}{costsFit2 object}
}
\description{
Plot traces for a subset of parameters using bayesplot
}
