% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_coal_times2}
\alias{simulate_coal_times2}
\title{Simulate coalescent process event times based on supplied trajectory}
\usage{
simulate_coal_times2(DF_in, samp_times, n_samp)
}
\arguments{
\item{DF_in}{Dataframe with columns: I - the trajectory, inf_rate - the per capita infection rate, t - time}

\item{samp_times}{Sampling times given as a time before present measured from simulation end point}

\item{n_samp}{number of samples taken at each sampling time}
}
\description{
Simulate coalescent process event times based on supplied trajectory
}
