% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_epi_nstrain}
\alias{simulate_epi_nstrain}
\title{Simulate tau-leaped trajectories from n-strain stochastic SIS}
\usage{
simulate_epi_nstrain(
  t0,
  tmax,
  nstep,
  n_strains,
  S_0,
  I_0,
  usage_func,
  beta_func,
  N_func,
  gamma_u,
  gamma_t
)
}
\arguments{
\item{t0}{start time}

\item{tmax}{end time}

\item{nstep}{number of tau-leaped steps}

\item{n_strains}{number of strains to simulate}

\item{S_0}{initial susceptible population}

\item{I_0}{initial infected populations, 1 per strain}

\item{usage_func}{abx of interest usage function}

\item{beta_func}{force of infection function}

\item{N_func}{population size function}

\item{gamma_u}{recovery rates when not treated with ABX of interest}

\item{gamma_t}{recovery rates when treated with ABX of interest}
}
\description{
Simulate tau-leaped trajectories from n-strain stochastic SIS
}
