% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_fast}
\alias{simulate_fast}
\title{Simulate SIS epidemic process}
\usage{
simulate_fast(
  u0,
  usage_vals,
  beta_vals,
  times,
  gamma_sus,
  gamma_res_u,
  gamma_res_t,
  dt,
  N
)
}
\arguments{
\item{u0}{Vector u0}

\item{usage_vals}{Vector of usage values}

\item{beta_vals}{Vector of beta values}

\item{times}{Vector of times}

\item{gamma_sus}{Recovery rate for sus}

\item{gamma_res_u}{Recovery rate for res untreated}

\item{gamma_res_t}{Recovery rate for res treated}

\item{dt}{dt}

\item{N}{N}
}
\description{
Simulate SIS epidemic process
}
