% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_fast_nstrain}
\alias{simulate_fast_nstrain}
\title{Simulate SIS epidemic process}
\usage{
simulate_fast_nstrain(
  n_strains,
  I0,
  S0,
  usage_vals,
  beta_vals,
  N_vals,
  gamma_u,
  gamma_t,
  times,
  dt
)
}
\description{
Simulate SIS epidemic process
}
