% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_phylos}
\alias{simulate_phylos}
\title{Simulate coalescent trees based on supplied trajectory}
\usage{
simulate_phylos(DF_in, samp_times, n_samp)
}
\arguments{
\item{DF_in}{Simulated trajectory generated by simulate_epi}

\item{samp_times}{Sampling times given as a time before present measured from simulation end point}

\item{n_samp}{number of samples taken at each sampling time}
}
\description{
Simulate coalescent trees based on supplied trajectory
}
