% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yearly_usg_stepfunc}
\alias{yearly_usg_stepfunc}
\title{Interpolate usage data using a step function}
\usage{
yearly_usg_stepfunc(usage, usage_ts)
}
\arguments{
\item{usage}{usage values}

\item{usage_ts}{usage time points}
}
\description{
Interpolate usage data using a step function
}
