% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs_fit2.R
\name{plot_rr_map}
\alias{plot_rr_map}
\title{Plot rt_sus - rt_res difference posterior probability heatmap}
\usage{
plot_rr_map(o, res_lineage_idx, n_breaks = 100, min_disp_prob = 0.5)
}
\arguments{
\item{o}{costsFit2 object}

\item{res_lineage_idx}{index of resistant lineage to display}

\item{n_breaks}{number of usage discretisation breaks}

\item{min_disp_prob}{minimum probability display cutoff}
}
\description{
f(U, c) = P[rt_sus - rt_res > c | u(t) = U]
}
