% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coal.R
\name{sim_coal}
\alias{sim_coal}
\title{Simulate coalescent process times using Poisson Process Thinning
Note: Ne function is interpolated as a step function}
\usage{
sim_coal(samp_t, n_samp, Ne_vals, Ne_times)
}
\arguments{
\item{samp_t}{a vector of sampling times}

\item{n_samp}{a vector containing the numbers of samples taken at corresponding sampling time.}

\item{Ne_vals}{a vector of Ne(t) values.}

\item{Ne_times}{a vector of time points corresponding to Ne(t) values.}
}
\description{
Simulate coalescent process times using Poisson Process Thinning
Note: Ne function is interpolated as a step function
}
