% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sim_coal_fast}
\alias{sim_coal_fast}
\title{Simulate coalescent tree using thinning}
\usage{
sim_coal_fast(samp_times, n_samp, Ne_vals, Ne_times, samp_total, dt, n)
}
\description{
Simulate coalescent tree using thinning
}
