% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_fast}
\alias{simulate_fast}
\title{Simulate SIS epidemic process}
\usage{
simulate_fast(
  u0,
  usage_vals,
  beta_vals,
  times,
  gamma_sus,
  gamma_res_u,
  gamma_res_t,
  dt,
  N
)
}
\description{
Simulate SIS epidemic process
}
